+++ sickbeard/providers/newznab.py
@@ -46,6 +46,11 @@ except ImportError:
     except ImportError:
         import xml.etree.ElementTree as etree
 
+try:
+    from collections import OrderedDict
+except ImportError:
+    from requests.compat import OrderedDict
+
 
 class NewznabConstants:
     SEARCH_TEXT = -100
@@ -120,6 +125,12 @@ class NewznabProvider(generic.NZBProvider):
         self._last_recent_search = None
         self._caps_last_updated = datetime.datetime.fromordinal(1)
         self.cache = NewznabCache(self)
+        # filters
+        if super(NewznabProvider, self).get_id() in ('nzbs_org',):
+            self.filter = []
+            if 'nzbs_org' == super(NewznabProvider, self).get_id():
+                self.may_filter = OrderedDict([
+                    ('so', ('scene only', False)), ('snn', ('scene not nuked', False))])
 
     @property
     def cat_ids(self):
@@ -710,6 +721,10 @@ class NewznabProvider(generic.NZBProvider):
                         and 'q' in params and not (any(x in params for x in ['season', 'ep'])):
                     request_params['t'] = 'search'
                 request_params.update(params)
+
+                if hasattr(self, 'filter'):
+                    if 'nzbs_org' == self.get_id():
+                        request_params['rls'] = ((0, 1)['so' in self.filter], 2)['snn' in self.filter]
 
                 # workaround a strange glitch
                 if sum(ord(i) for i in self.get_id()) in [383] and 5 == 14 - request_params['maxage']: